apps/modal-backend/requirements.txt
modal>=0.63.0
fastapi>=0.104.0
pydantic>=2.4.0
pillow>=10.0.0
torch>=2.0.0
transformers>=4.35.0
diffusers>=0.21.0
accelerate>=0.24.0
numpy>=1.24.0
aiofiles>=23.2.0
python-multipart>=0.0.6
uvicorn>=0.24.0